requirements.txt
asgiref~=3.8.1
channels~=4.1.0
Babel~=2.15.0
certifi~=2024.7.4
charset-normalizer~=3.3.2
Django~=5.0.7
django-modeladmin-reorder~=0.3.1
idna~=3.7
pytz~=2024.1
requests~=2.32.3
sqlparse~=0.5.1
urllib3~=2.2.2
python-dotenv~=1.0.1
envparse~=0.2.0
telethon~=1.36.0
tgcrypto~=1.2.5
rsa~=4.9
django-solo~=2.3.0
